[Edited]: data type ของ table field ใน database เปลี่ยนจาก BIGINT ไปเป็น int

+++ tearup/init.sql
@@ -25,7 +25,7 @@ CREATE TABLE products (
 INSERT INTO products (id,product_name,product_brand,quantity,product_price,image_url) VALUE (2,"43 Piece dinner Set","CoolKidz",10,12.95,"/43_Piece_dinner_Set.png");
 
 CREATE TABLE orders (
-    id BIGINT AUTO_INCREMENT,
+    id int AUTO_INCREMENT,
     total_price double,
     transaction_id varchar(255) DEFAULT '',
     completed smallint(1) DEFAULT 0,
@@ -36,7 +36,7 @@ CREATE TABLE orders (
 );
 
 CREATE TABLE order_product (
-    order_id BIGINT,
+    order_id int,
     product_id int,
     quantity int,
     product_price double
@@ -44,7 +44,7 @@ CREATE TABLE order_product (
 
 CREATE TABLE shipping (
     id int AUTO_INCREMENT,
-    order_id BIGINT,
+    order_id int,
     address varchar(255),
     sub_district varchar(255),
     district varchar(255),
